Rename and change location of Navigation
